move inoput buttons into top level component file
