+++ packages/pretzelai-extension/src/components/InputComponent.tsx
@@ -13,9 +13,108 @@ import { EditorView, keymap, placeholder } from '@codemirror/view';
 import { markdown } from '@codemirror/lang-markdown';
 import { defaultHighlightStyle, syntaxHighlighting } from '@codemirror/language';
 import { history, historyKeymap, insertNewlineAndIndent } from '@codemirror/commands';
-import { PromptHistoryButton, RemoveButton, SubmitButton } from './prompt-box-buttons';
 import posthog from 'posthog-js';
 import { FixedSizeStack } from '../utils';
+
+import { LabIcon } from '@jupyterlab/ui-components';
+import promptHistorySvg from '../../style/icons/prompt-history.svg';
+import '../../style/base.css';
+
+interface ISubmitButtonProps {
+  handleClick: () => void;
+  isDisabled: boolean;
+}
+
+const SubmitButton: React.FC<ISubmitButtonProps> = ({ handleClick, isDisabled }) => {
+  const [showTooltip, setShowTooltip] = useState(false);
+
+  return (
+    <div className="submit-button-container">
+      <button
+        className="pretzelInputSubmitButton"
+        onClick={handleClick}
+        disabled={isDisabled}
+        onMouseEnter={() => setShowTooltip(true)}
+        onMouseLeave={() => setShowTooltip(false)}
+        title="Submit ↵"
+      >
+        Submit <span style={{ fontSize: '0.8em' }}>↵</span>
+      </button>
+      {showTooltip && (
+        <div className="tooltip">
+          Send prompt to AI for completion <strong>(Enter)</strong>
+        </div>
+      )}
+    </div>
+  );
+};
+
+interface IRemoveButtonProps {
+  handleClick: () => void;
+}
+
+const RemoveButton: React.FC<IRemoveButtonProps> = ({ handleClick }) => {
+  const [showTooltip, setShowTooltip] = useState(false);
+  const isMac = /Mac/i.test(navigator.userAgent);
+  const keyCombination = isMac ? 'Cmd + K' : 'Ctrl + K';
+  const shortcut = isMac ? '⌘K' : '^K';
+
+  return (
+    <div className="remove-button-container">
+      <button
+        className="remove-button"
+        onClick={handleClick}
+        onMouseEnter={() => setShowTooltip(true)}
+        onMouseLeave={() => setShowTooltip(false)}
+        title={`Remove ${shortcut}`}
+      >
+        Remove <span style={{ fontSize: '0.8em' }}>{shortcut}</span>
+      </button>
+      {showTooltip && (
+        <div className="tooltip">
+          Remove the AI prompt box <strong>({keyCombination})</strong>
+        </div>
+      )}
+    </div>
+  );
+};
+
+const encodedSvgStr = encodeURIComponent(promptHistorySvg);
+
+const promptHistoryIcon = new LabIcon({
+  name: 'pretzelai::prompt-history',
+  svgstr: encodedSvgStr
+});
+
+const PromptHistoryButton: React.FC<{
+  handleClick: (promptHistoryIndex: number) => void;
+  promptHistoryIndex: number;
+}> = ({ handleClick, promptHistoryIndex }) => {
+  const [showTooltip, setShowTooltip] = useState(false);
+  const isMac = /Mac/i.test(navigator.userAgent);
+  const keyCombination = isMac ? 'Cmd + Shift + H' : 'Ctrl + Shift + H';
+
+  return (
+    <div className="prompt-history-button-container">
+      <button
+        className="prompt-history-button"
+        title="Prompt History"
+        onClick={() => {
+          handleClick(promptHistoryIndex);
+        }}
+        onMouseEnter={() => setShowTooltip(true)}
+        onMouseLeave={() => setShowTooltip(false)}
+      >
+        <promptHistoryIcon.react tag="span" className="jp-Icon jp-Icon-20" />
+      </button>
+      {showTooltip && (
+        <div className="tooltip">
+          Populate with last prompt <strong>({keyCombination})</strong>
+        </div>
+      )}
+    </div>
+  );
+};
 
 interface IInputComponentProps {
   isAIEnabled: boolean;
@@ -27,9 +126,8 @@ interface IInputComponentProps {
   promptHistoryStack: FixedSizeStack<string>;
   setInputView: (view: EditorView) => void;
   initialPrompt?: string;
-  activeCell: any; // Add this prop
+  activeCell: any;
 }
-
 const InputComponent: React.FC<IInputComponentProps> = ({
   isAIEnabled,
   placeholderEnabled,
@@ -40,7 +138,7 @@ const InputComponent: React.FC<IInputComponentProps> = ({
   promptHistoryStack,
   setInputView,
   initialPrompt = '',
-  activeCell // Add this prop
+  activeCell
 }) => {
   const inputFieldRef = useRef<HTMLDivElement>(null);
   const inputViewRef = useRef<EditorView | null>(null);
